iterate: refactoring
